pyproject.toml
[project]
name = "crosnetik-be"
version = "0.1.0"
description = ""
authors = [
    {name = "Cybertronian123",email = "atrayanmukherjee0248@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn[standard] (>=0.41.0,<0.42.0)",
    "motor (>=3.7.1,<4.0.0)",
    "beanie (>=2.0.1,<3.0.0)",
    "pymongo (>=4.16.0,<5.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "passlib[argon2] (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "flake8 (>=7.3.0,<8.0.0)",
    "black (>=26.3.1,<27.0.0)",
    "isort (>=8.0.1,<9.0.0)",
    "pytest (>=9.0.2,<10.0.0)"
]
